Refactor `installPrettier`
