+++ tests/config/install-prettier.js
@@ -1,36 +1,38 @@
 "use strict";
 
 const path = require("path");
-const shell = require("shelljs");
+const fs = require("fs");
+const execa = require("execa");
 const tempy = require("tempy");
-
-shell.config.fatal = true;
 
 const client = process.env.NPM_CLIENT || "yarn";
 
 module.exports = (packageDir) => {
   const tmpDir = tempy.directory();
-  const file = shell.exec("npm pack", { cwd: packageDir }).stdout.trim();
-  shell.mv(path.join(packageDir, file), tmpDir);
+  let file = execa.sync("npm", ["pack"], { cwd: packageDir }).stdout.trim();
+  file = path.join(packageDir, file);
   const tarPath = path.join(tmpDir, file);
+  fs.copyFileSync(file, tarPath);
+  fs.unlinkSync(file);
 
-  shell.exec(`${client} init -y`, { cwd: tmpDir, silent: true });
-  let installCommand = "";
+  execa.sync(client, ["init", "-y"], { cwd: tmpDir });
+
+  let installArguments = [];
   switch (client) {
     case "npm":
       // npm fails when engine requirement only with `--engine-strict`
-      installCommand = `npm install "${tarPath}" --engine-strict`;
+      installArguments = ["install", tarPath, "--engine-strict"];
       break;
     case "pnpm":
       // Note: current pnpm can't work with `--engine-strict` and engineStrict setting in `.npmrc`
-      installCommand = `pnpm add "${tarPath}"`;
+      installArguments = ["add", tarPath, "--engine-strict"];
       break;
     default:
       // yarn fails when engine requirement not compatible by default
-      installCommand = `yarn add "${tarPath}"`;
+      installArguments = ["add", tarPath];
   }
 
-  shell.exec(installCommand, { cwd: tmpDir });
+  execa.sync(client, installArguments, { cwd: tmpDir });
 
   return path.join(tmpDir, "node_modules/prettier");
 };
